Added a "RemoteWrite" section to e2e tests [NOT-WORKING].
I though I will find some errors in the logs of prometheus's pod but
this is not the case, there is no mention of TLS handshake errors.

Conclusion:
No ServiceMonitor is created therefore there is no scraped data so
nothing is send via RemoteWrite and we don't get any error.
I need to make sure I create ServiceMonitor CR as well as Prometheus.

Also I need to change RemoteWrite.URL to the app's ingress instead of
the app's service because only the ingress is enforcing TLS.

Signed-off-by: Yoni Bettan <ybettan@redhat.com>